Move fieldContext to utils/ and useAppForm to hooks/
Better separation of concerns: context setup lives with utilities,
the form hook lives with other hooks.
