+++ packages/client/src/components/formFields/index.ts
@@ -3,4 +3,4 @@ export { InputField } from "./InputField";
 export { NumberField } from "./NumberField";
 export { RadioGroupField } from "./RadioGroupField";
 export { TextareaField } from "./TextareaField";
-export { useAppForm } from "./formHook";
+export { useAppForm } from "@/hooks/useAppForm";
