+++ packages/client/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./fetchFunctions";
+export * from "./fieldContext";
 export * from "./formHasChanges";
 export * from "./makePercentageComplete";
